mypublish to root
