+++ project/BldBridgeScoreKeeper.scala
@@ -407,112 +407,6 @@ object BldBridgeScoreKeeper {
       disttests := Def
         .sequential(integrationtests in Distribution, moretests in Distribution)
         .value,
-      publishdir := {
-        // returns an Option[File]
-
-        val log = streams.value.log
-
-        import java.io.File
-        sys.props("user.home") match {
-          case homedir if (homedir != null) =>
-            val configfile = new File(homedir, "bridgescorer/config.properties")
-            if (configfile.exists()) {
-              import java.util.Properties
-              import java.io.InputStreamReader
-              import java.io.FileInputStream
-              val props = new Properties
-              props.load(
-                new InputStreamReader(new FileInputStream(configfile), "UTF8")
-              )
-              val dd = props.getProperty("DistributionDirectory")
-              if (dd != null) {
-                import java.io.File
-
-                val distdir = dd.replace('\\', '/')
-                val f = new File(distdir)
-                if (f.isDirectory()) {
-                  log.info("Publishing to " + f)
-                  Some(f)
-                } else {
-                  throw new RuntimeException(
-                    "DistributionDirectory directory does not exist: " + f
-                  )
-                  None
-                }
-              } else {
-                throw new RuntimeException(
-                  "DistributionDirectory property does not exist in file ~/bridgescorer/config.properties"
-                )
-                None
-              }
-            } else {
-              throw new RuntimeException(
-                "file ~/bridgescorer/config.properties does not exist"
-              )
-              None
-            }
-          case _ =>
-            throw new RuntimeException("Home directory not set")
-            None
-        }
-      },
-      mypublishcopy := {
-
-        val log = streams.value.log
-
-        val dd = publishdir.value
-        dd match {
-          case Some(distdir) =>
-            import java.nio.file.Path
-            import java.nio.file.StandardCopyOption
-            import java.nio.file.Files
-
-            log.info("Publishing to " + distdir)
-
-            val targetdir = (crossTarget in Compile).value
-            val assemblyjar = (assemblyJarName in assembly).value
-            val testjar = (assemblyJarName in (Test, assembly)).value
-
-            val sourceassemblyjar = new File(targetdir, assemblyjar)
-            val targetassemblyjar = new File(distdir, assemblyjar)
-            val sourcetestjar = new File(targetdir, testjar)
-            val targettestjar = new File(distdir, testjar)
-
-            IO.listFiles(distdir, GlobFilter("*.jar")).foreach { jar =>
-              {
-                log.info("Moving jar to save: " + jar)
-                IO.move(
-                  jar,
-                  new File(new File(jar.getParentFile, "save"), jar.getName)
-                )
-              }
-            }
-
-            log.info("Publishing " + assemblyjar + " to " + distdir)
-            Files.copy(
-              sourceassemblyjar.toPath,
-              targetassemblyjar.toPath,
-              StandardCopyOption.REPLACE_EXISTING
-            )
-            log.info("Publishing " + testjar + " to " + distdir)
-            Files.copy(
-              sourcetestjar.toPath,
-              targettestjar.toPath,
-              StandardCopyOption.REPLACE_EXISTING
-            )
-
-            log.info("Published to " + distdir)
-          case None =>
-            throw new RuntimeException("DistributionDirectory is not set")
-        }
-
-      },
-      mypublish := Def
-        .sequential(
-          disttests in Distribution,
-          mypublishcopy in Distribution
-        )
-        .value
     )
 
     def cleanCacheDir( stream: sbt.Keys.TaskStreams ) = {
